requirements.txt
aioredis==1.3.1
asgiref==3.5.2
asyncio==3.4.3
aiohttp==3.8.0
async-timeout==4.0.2
attrs==21.2.0
autobahn==22.5.1
Automat==20.2.0
beautifulsoup4==4.9.3
boto3==1.24.30
celery==5.2.7
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.4.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssutils==2.2.0
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.51.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-dotenv==1.4.2
django-environ==0.4.5
django-filter==2.4.0
django-inlinecss==0.3.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
django-on-heroku==1.1.2
django-celery-beat==2.3.0
django-redis==5.2.0
django-storages==1.12.3
drf-extra-fields==3.4.0
drf-yasg==1.20.0
deepface==0.0.75
future==0.18.2
futures3==1.0.0
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
opencv-python==4.6.0.66
packaging==21.0
Pillow==8.1.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pynliner==0.8.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2022.1
redis==3.4.1
requests==2.25.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
service-identity==21.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
zipp==3.4.1
zope.interface==5.4.0
